Multiple small changes mostly related to code quality

+++ src/main/java/ch/niceideas/eskimo/services/TerminalServiceImpl.java
@@ -104,7 +104,7 @@ public class TerminalServiceImpl implements TerminalService {
                         }
                     }
                 }
-            }, idleTimeoutSeconds / (long) 10 * 1000, idleTimeoutSeconds / (long) 10 * 1000);
+            }, idleTimeoutSeconds / 10 * 1000, idleTimeoutSeconds / 10 * 1000);
     }
 
     @PreDestroy
